move react components out of /components and into /src
